# Use session identity instead of tokens for worker and kernel