--- a/apps/server/lib/http/b64.ts
+++ b/apps/server/lib/http/b64.ts
@@ -1,0 +1,3 @@
+export const b64decode = (str: string): string => {
+	return Buffer.from(str, 'base64').toString().trim();
+};