--- a/apps/server/lib/plugins.ts
+++ b/apps/server/lib/plugins.ts
@@ -16,5 +16,5 @@ export function getPlugins(): PluginDefinition[] {
 		outreachPlugin(),
 		frontPlugin(),
 		balenaApiPlugin(),
-	];
+	] as any;
 }