--- a/apps/server/tsconfig.json
+++ b/apps/server/tsconfig.json
@@ -22,6 +22,10 @@
   },
   "include": [
     "lib/**/*.ts",
+    "@types/**/*.ts",
     "test/**/*.ts"
-  ]
+  ],
+  "ts-node": {
+    "files": true
+  }
 }