--- a/apps/ui/package.json
+++ b/apps/ui/package.json
@@ -23,7 +23,7 @@
   "author": "Balena.io. <hello@balena.io>",
   "license": "AGPL-3.0",
   "dependencies": {
-    "@balena/jellyfish-client-sdk": "^12.0.0",
+    "@balena/jellyfish-client-sdk": "^13.0.0",
     "@balena/jellyfish-environment": "^12.2.2",
     "@sentry/browser": "^6.19.7",
     "@use-it/event-listener": "^0.1.7",