… apps/server/lib/http/auth.ts
… apps/server/lib/socket/index.ts
… apps/server/package.json
… apps/server/test/integration/helpers.ts
… package.json
… test/e2e/server/notifications.spec.js
… test/e2e/server/registry.spec.js
… test/e2e/server/sessions.spec.js
… test/e2e/ui/livechat/index.spec.js
… test/e2e/ui/livechat/macros.js