This change brings in the [update to autumndb](https://github.com/product-os/autumndb/pull/1363) that removes direct session handling.
Instead of AutumnDB handling session tokens, it only deals with the identity/actor making requests. Authorization of users is now hoisted up to the route layer at the top of the call stack, where it can eventually be delegated to an external service, such as Ory Kratos.
This has the effect of only retrieving the actor once at the top of the call tree, rather than in every interaction with the autumnDB kernel interface, which brings a fairly hefty performance boost.

Change-type: major
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>

***

**Please remember to write tests for your changes**. We aim to automatically
deploy `master` to production, and we can't safely do this without a solid
test suite!