@LucianBuzzo 2022-06-23
Running some simple benchmarking locally, this change brings a ~20% performance increase when creating a thread.

I used the following bench script:
```
const { getSdk } = require('@balena/jellyfish-client-sdk');
const bench = require('nanobench');

const setup = async () => {
	const sdk = getSdk({
		apiPrefix: 'api/v2',
		apiUrl: 'http://localhost:8000',
	});

	await sdk.auth.login({
		username: 'jellyfish',
		password: 'jellyfish',
	});
	return sdk;
};

bench('create thread 100 times', async function(b) {
	const times = 100;
	const sdk = await setup();

	b.start();

	for (var i = 0; i < times; i++) {
		// Create a thread with the SDK
		await sdk.card.create({
			type: 'thread@1.0.0',
		});
	}

	b.end();
});
```

On master this gives the following result:
```
ok ~26 s (25 s + 976374500 ns)
```

On this branch I get:
```
ok ~20 s (20 s + 148205800 ns)
```