@LucianBuzzo 2022-06-23
@joshbwlng I would have used hyperfine here, but I had some trouble installing it due to missing deps on my machine