@joshbwlng 2022-06-23
@LucianBuzzo We could make the switch to `nanobench` instead of that works better for everyone. I'd rather have one tool instead of multiple. And great work by the way, another ~20% performance boost is amazing! 👍