@LucianBuzzo 2022-06-24
@joshbwlng I could probably get hyperfine working with a bit of effort, and it's language agnostic which is nice - maybe its a conversation for another time.
The performance gains are very nice - now I just need to get these e2e UI tests passing 😮‍💨