@LucianBuzzo 2022-06-24
@karaxuna This PR is consistently failing on this livechat test:
```
[e2e-ui]   1) test/e2e/ui/livechat/index.spec.js:172:1 › Create conversation page ===========================
[e2e-ui] 
[e2e-ui]     Error: No notifications found
[e2e-ui] 
[e2e-ui]       213 |
[e2e-ui]       214 | 	// Should receive notification
[e2e-ui]     > 215 | 	const [ notification ] = await macros.waitForNotifications(page, 1)
[e2e-ui]           | 	                         ^
[e2e-ui]       216 | 	expect(notification.links['is attached to'][0].id).toEqual(response.id)
[e2e-ui]       217 |
[e2e-ui]       218 | 	// Should display support agent's username
[e2e-ui] 
[e2e-ui]         at /usr/src/jellyfish/test/e2e/ui/livechat/macros.js:190:10
[e2e-ui]         at Object.exports.retry (/usr/src/jellyfish/test/e2e/ui/livechat/macros.js:12:18)
[e2e-ui]         at /usr/src/jellyfish/test/e2e/ui/livechat/index.spec.js:215:27
[e2e-ui] 
```

Any ideas how I can debug this? Everything seems to work fine locally 🤔