@karaxuna 2022-06-25
This notifications are set here: https://github.com/product-os/jellyfish/blob/master/apps/ui/lib/components/Livechat.tsx#L32

I can take a look this weekends.