@LucianBuzzo 2022-06-27
@karaxuna It's been really hard to get an accurate local setup working. I got something going, but the `product` value wasn't set - when I made it optional everything worked as expected.