@LucianBuzzo 2022-06-27
Ok, I found the issue. When the worker generates a new notification, it first creates session on behalf of the user that created the subscription. Unfortunately this session is being created using the session of the *triggering* user, not the privileged user, resulting in a session where the actor doesn't have any roles, because regular users can't see the roles of other users. 🤦